Lecture 03 - Part 2 Code

+++ 03-Part2/point.h
@@ -12,7 +12,7 @@
 
 #include <iostream>
 
-typedef float COORDINATE;
+typedef double COORDINATE;
 
 
 class point{
